\name{fsts}
\docType{data}
\alias{fsts}
\title{Example fst calculations from a genepop file.}
\description{
   Example fst calculations from a genepop file.
   The original data were generated by using a numerical analysis with Nm = 10, 75 demes, and 5 population samples taken. No selection was imposed.
	The fsts were calculated using calc.actual.fst(gpop)
   This file contains a dataframe with 2000 columns and 3 rows. The first column is the Locus ID, the second column is the Ht for that locus, and the third column is the Fst for that locus.
}
\usage{fsts}
\format{data.frame}
\source{Generated by numerical analysis}
\references{
   See Flanagan & Jones
}