\name{genepop.example}
\docType{data}
\alias{genepop.example}
\title{Example genepop file from numerical simulations}
\description{
   Example genepop file from numerical simulations.
   It was generated by using a numerical analysis with Nm = 10, 75 demes, and 5 population samples taken. No selection was imposed.
   This file contains a dataframe with 2002 columns and 250 rows. The first two columns are the population name and the individual name. The remaining columns are genotypes for each locus (one column per locus). Each row is an individual.
}
\usage{gpop}
\format{data.frame}
\source{Generated by numerical analysis}
\references{
   See Flanagan & Jones
}